swiping on phone app

+++ superb_ock/models.py
@@ -11,7 +11,7 @@ class GolfCourse(models.Model):
     country = models.CharField(max_length=20,null=True)
     tees = models.CharField(max_length=20,default='White')
     slope_rating = models.IntegerField(default=113)
-    course_rating = models.DecimalField(decimal_places=1,max_digits=3,default=72)
+    course_rating = models.FloatField(default=72)
     par = models.IntegerField(default=72)
     slug = models.SlugField(default="", null=False)
